requirements.txt
anthropic>=0.25.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pypdf>=4.0.0
python-docx>=1.1.0